refactor: ChangeReporter
