+++ packages/cli/resources/change-report/openapi-diff/details.hbs
@@ -1,5 +1,5 @@
 {{#*inline "api-change"}}
-{{title}} {{#each sourceSpecEntityDetails}}`{{location}}`{{/each}} {{#each destinationSpecEntityDetails}}`{{location}}`{{/each}}
+{{title}} `{{location}}`
 {{~/inline}}
 {{#if openapiDiff.differenceCount}}
 <h2 id="openapi-changes">🔄 API changes</h2>
