+++ packages/cli/tests/unit/diffArchive/fixtureData/addDeleteChangeRemoveDiff.txt
