+++ packages/cli/tests/unit/diffArchive/fixtureData/exampleDiff.txt
