+++ packages/cli/tests/unit/diffArchive/fixtureData/expectedDiff.txt
