+++ packages/cli/tests/unit/fixtures/compare/.appmap/change-report/all-succeeded/change-report.json
@@ -752,7 +752,7 @@
         "code": "response.status-code.add",
         "destinationSpecEntityDetails": [
           {
-            "location": "302 GET /users/{id}",
+            "location": "paths./users/{id}.get.responses.302",
             "value": {
               "content": {
                 "text/html": {}
